slide_analyzer_backend/requirements.txt
fastapi
uvicorn
openai
python-dotenv
pdf2image
python-pptx
Pillow
pydantic[email]
google-api-python-client
google-auth
fitz
frontend
tools